doc: corrected HTML attributes in README.md
The `align` attribute has been deprecated for ages, and shouldn't be used anymore.

The `width` and `height` attributes are only valid in a few tags (such as `img`) and they should be plain integers corresponding to the CSS pixel dimensions.

The dimensions listed here are based on the adjusted image I proposed in #1004 and https://github.com/ZQuestClassic/commons/pull/1 . Of course we can fine-tune those dimensions if the image changes.

* https://html.spec.whatwg.org/#dimension-attributes
* https://www.w3.org/TR/html5-diff/#obsolete-attributes
* https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes